% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building_Rooftop_Shape_Analysis_Lidar.R
\name{clipping_lidar_buildings_opPlot}
\alias{clipping_lidar_buildings_opPlot}
\title{clipping_lidar_buildings_opPlot Function}
\usage{
clipping_lidar_buildings_opPlot(path_lidar_file, path_shp_buildings,
  path_folder_save = "", make_plot = TRUE, name_id_field = "")
}
\arguments{
\item{path_lidar_file}{This is the path which contains the main Lidar file (.las)}

\item{path_shp_buildings}{This is the path that contains the shapefile file with each of the roofs of the buildings.}

\item{path_folder_save}{This is the directory path where the user wishes to save the cut lidar files for each of the buildings (roofs), This is the directory path where the user wishes to save the cut lidar files for each of the buildings (roofs), if the parameter is empty the results will be stored in the tempdir folder}

\item{make_plot}{This is a Boolean option which evaluates whether or not the user wishes to plot the lidar files generated after the clip operation with the buildings.}

\item{name_id_field}{This is an optional parameter where the user can specify which field identifies the shapefile of the buildings. The value of this field will be the identifier for each of the generated files.}
}
\value{
list_new_las It is a list which contains each of the .las files generated by cutting the main lidar file with the roofs from the shapefile
}
\description{
This function is responsible for making the clip of a lidar file for each of the
potential roofs to develop hubs for air taxis. The function has the option of
plotting each of the terraces in 3D in order to have a better visualization of
the cloud of 3D points.
This function returns a list with each of the Lidar files (.las),
which are the results of the clip made to each of the buildings. The user can specify
a specific route or if not assigned, the results will be stored in the tempdir folder
}
\examples{
path_las <- system.file("extdata/new_lidar.las", package = "RooftopsHubsAnalyisisLidar")
sample_rooftops_print <- system.file("extdata/sample_roofs.shp", package = "RooftopsHubsAnalyisisLidar")
list_las_elements <- clipping_lidar_buildings_opPlot(path_las, sample_rooftops_print)
}
\keyword{buildings,}
\keyword{clip}
\keyword{lidar,}
\keyword{plot,}
\keyword{rooftops,}
\keyword{shapefile,}
