% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building_Rooftop_Shape_Analysis_Lidar.R
\name{dist_heigh_boxplots_lidar}
\alias{dist_heigh_boxplots_lidar}
\title{dist_heigh_boxplots_lidar Function}
\usage{
dist_heigh_boxplots_lidar(list_elements_lidar)
}
\arguments{
\item{list_elements_lidar}{List with the generated lidar files to perform the boxplot of each one of them}
}
\description{
Function to plot the boxplots of the heights of each building
}
\examples{
library(ggplot2)
path_folder_las <- system.file("extdata/sample_las", package = "RooftopsHubsAnalyisisLidar")
list_las <- base::list.files(path_folder_las,full.names = TRUE)
dist_heigh_boxplots_lidar(list_las)
}
