% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Building_Rooftop_Shape_Analysis_Lidar.R
\name{verify_id_field}
\alias{verify_id_field}
\title{verify_id_field Function}
\usage{
verify_id_field(path_shp_buildings, name_id_field)
}
\arguments{
\item{path_shp_buildings}{Path of the shapefile to perform the validation}

\item{name_id_field}{Name of the field to verify if it exits within the shapefile}
}
\description{
Function to verify if the field provided by the user exists or not.
}
