% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Available_Areas_Noise_Analysis_Rooftops.R
\name{verify_id_field_shapefile}
\alias{verify_id_field_shapefile}
\title{verify_id_field_shapefile Function}
\usage{
verify_id_field_shapefile(shp_buildings, name_id_field)
}
\arguments{
\item{shp_buildings}{Shapefile to perform the validation}

\item{name_id_field}{name_id_field Name of the field to verify if it exits within the shapefile}
}
\description{
Function to verify if the field provided by the user exists or not
}
