% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Available_Areas_Noise_Analysis_Rooftops.R
\name{verify_projection_epsg_shp}
\alias{verify_projection_epsg_shp}
\title{verify_projection_epsg_shp Function}
\usage{
verify_projection_epsg_shp(shp_entity)
}
\arguments{
\item{shp_entity}{Shapefile of interest to perform the validation}
}
\description{
Function that perform the validation of the projection and epsg of a shapefile
}
