% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Available_Areas_Noise_Analysis_Rooftops.R
\name{verify_projection_raster}
\alias{verify_projection_raster}
\title{verify_projection_raster Function}
\usage{
verify_projection_raster(raster_entity)
}
\arguments{
\item{raster_entity}{Noise map in raster format}
}
\description{
Function that perform the validation of the projection of a raster
}
